Use C99 number types. Tidy up RISC OS application skeleton.
* !ArcEm/!Boot,feb - Give !ArcEm a !Boot file
* !ArcEm/!Help - Rewritten help file to reflect current status
* !ArcEm/!Run,feb - Updated !Run file
* c99.h - New header to provide a common place to implement fallbacks for any C99 functionality that may be missing on specific platforms
* Any files not listed above - Updated to use C99 number types

+++ X/ControlPane.c
@@ -56,7 +56,7 @@ static void DoLED(const char *Text, int On, int ty, int lx)
 
 /*----------------------------------------------------------------------------*/
 
-static void draw_keyboard_leds(unsigned int leds)
+static void draw_keyboard_leds(uint8_t leds)
 {
     DoLED("Caps Lock",   leds & KBD_LED_CAPSLOCK, LEDTOPS, 0);
     DoLED("Num Lock",    leds & KBD_LED_NUMLOCK, LEDTOPS, 90);
